test script in place (2 client and 1 server) +++ client registration to the service ok

+++ P2PUIClient/P2PClient.h
@@ -25,7 +25,7 @@ struct P2PClient
   void Stop();
 
 private:
-  const uint8_t mUID;
+  const short mUID;
   IRequester& mServerRequester;
   ISubscriber& mServerSubscriber;
   IReplier& mPeerReplier;
